move utils to their own dir
