+++ src/commands/music/shuffle.ts
@@ -6,7 +6,7 @@ import {
 } from '@sapphire/framework';
 import type { CommandInteraction } from 'discord.js';
 import { container } from '@sapphire/framework';
-import type { Song } from '../../lib/queue/Song';
+import type { Song } from '../../lib/utils/queue/Song';
 
 @ApplyOptions<CommandOptions>({
   name: 'shuffle',
