+++ src/lib/utils/trivia/utilFunctions.ts
